# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Chainsaw Manual Training App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /waiver:
    get:
      operationId: getWaiver
      tags: [waiver]
      summary: Get waiver status for current user
      parameters:
        - name: deviceId
          in: header
          required: true
          schema:
            type: string
        - name: activationCode
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Waiver status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaiverStatus"

    post:
      operationId: signWaiver
      tags: [waiver]
      summary: Sign the digital waiver with signature data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WaiverInput"
      responses:
        "200":
          description: Waiver signed and PDF created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    WaiverStatus:
      type: object
      required: [signed]
      properties:
        signed:
          type: boolean
        signedAt:
          type: ["string", "null"]
        pdfUrl:
          type: ["string", "null"]
        fullName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        signatureData:
          type: ["string", "null"]
        clausesSnapshot:
          type: ["string", "null"]
    WaiverInput:
      type: object
      required: [deviceId, activationCode, signatureData, agreedToTerms]
      properties:
        deviceId:
          type: string
        activationCode:
          type: string
        signatureData:
          type: string
        agreedToTerms:
          type: boolean
        clausesSnapshot:
          type: string
          description: JSON snapshot of clause titles and texts at the moment of signing
...